Use Ubuntu's own JDK
The JDK used by the parent Tomcat image is not built but installed from
binaries. The built by Ubunu jdk is faster and more stable in terms of
memory usage compared to the one shipped by Tomcat.

There is also no longer the need to install the jprgturbo plugin as JDK
is now using it internally by default to for compression.